+++ src/core/cache/CacheService.ts
@@ -67,7 +67,9 @@ export class CacheService {
     public async updateCaches(events: ObjectUpdatedEventData[]): Promise<void> {
         for (const event of events) {
             LoggingService.getInstance().debug('Backend Notification: ' + JSON.stringify(event));
-            await this.deleteKeys(event.Namespace);
+            if (!event.Namespace.startsWith(KIXObjectType.TRANSLATION)) {
+                await this.deleteKeys(event.Namespace);
+            }
         }
     }
